examples/boid: Structure in agents.py and model.py
