+++ examples/basic/boid_flockers/app.py
@@ -1,4 +1,4 @@
-from boid_flockers.model import BoidFlockers
+from model import BoidFlockers
 from mesa.visualization import SolaraViz, make_space_matplotlib
 from mesa.visualization import Slider
